requirements.txt
pyjwt
cryptography
requests
coinbase
attrs
idna_ssl
aiohttp
async_timeout
charset_normalizer
async_timeout
yarl
psycopg2
crypto_data
dash
sqlalchemy
plotly.express
dash-bootstrap-components
